railway_passenger_monitoring/requirements.txt
face_recognition==1.3.0
opencv-python==4.7.0.72
pymongo==4.3.3
cryptography==41.0.3
python-dotenv==1.0.0